Move loading process to the fragment itself
