+++ app/src/main/java/org/wikipedia/feed/suggestededits/SuggestedEditsCard.kt
@@ -3,13 +3,12 @@ package org.wikipedia.feed.suggestededits
 import org.wikipedia.R
 import org.wikipedia.WikipediaApp
 import org.wikipedia.dataclient.WikiSite
-import org.wikipedia.dataclient.mwapi.MwQueryPage
+import org.wikipedia.descriptions.DescriptionEditActivity
 import org.wikipedia.feed.model.CardType
 import org.wikipedia.feed.model.WikiSiteCard
 import org.wikipedia.util.DateUtil
 
-class SuggestedEditsCard(val summaryList: List<SuggestedEditsFeedClient.SuggestedEditsSummary>?,
-                         val imageTagsPage: MwQueryPage?,
+class SuggestedEditsCard(val cardTypes: List<DescriptionEditActivity.Action>,
                          wiki: WikiSite,
                          val age: Int) : WikiSiteCard(wiki) {
